Supporting server client plugin instead of the service export artifect, all the logic is in the server module now

+++ scripts/run.bat
@@ -3,7 +3,15 @@
 LOGS_DIR=logs/
 CONF_DIR=conf/
 
+echo "Staring DroneGCS"
 mkdir $LOGS_DIR
 
-sudo ln -s /dev/ttyACM0 /dev/ttyS85
+echo "Verify USB device"
+if [ ! -e /dev/ttyS85 ]; then
+	echo "Creating device link" 
+	sudo ln -s /dev/ttyACM0 /dev/ttyS85
+fi
+echo "Device link created" 
+
+echo "Starting dashboard" 
 java -Djava.library.path=../lib -cp /usr/local/jdk1.8.0_121/jre/lib/ext/jfxrt.jar:'*'  com.dronegcs.console.controllers.DroneLaunch $LOGS_DIR $CONF_DIR> $LOGS_DIR/backlog.elg
